SDK-1470: Move classes into yoti_sandbox directory and add TokenRequestBuilder
