+++ lib/yoti_sandbox.rb
@@ -1,4 +1,1 @@
-require_relative 'profile/anchor'
-require_relative 'profile/attribute'
-require_relative 'profile/profile'
-require_relative 'profile/sandbox_client'
+require_relative 'yoti_sandbox/profile'
