+++ lib/yoti_sandbox/profile.rb
@@ -1,0 +1,4 @@
+require_relative 'profile/request/anchor'
+require_relative 'profile/request/attribute'
+require_relative 'profile/request/token_request'
+require_relative 'profile/client'
