+++ spec/spec_helper.rb
@@ -6,7 +6,7 @@ $LOAD_PATH.unshift File.expand_path('../lib', __dir__)
 
 require 'webmock/rspec'
 
-require 'yoti_sandbox'
+require 'yoti_sandbox/profile'
 
 RSpec.configure do |config|
   config.before do
