+++ spec/yoti_sandbox/profile/request/attribute_spec.rb
@@ -2,7 +2,7 @@
 
 require 'spec_helper'
 
-describe 'Sandbox::Attribute' do
+describe 'Sandbox::Profile::Request::Attribute' do
   let :name do
     'name'
   end
@@ -14,19 +14,19 @@ describe 'Sandbox::Attribute' do
   end
   describe '.with_anchor' do
     let :anchor do
-      Sandbox::Anchor.source('anchorValue', timestamp: Time.utc(2008, 7, 6, 0, 0))
+      Sandbox::Profile::Request::Anchor.source('anchorValue', timestamp: Time.utc(2008, 7, 6, 0, 0))
     end
     let :attribute do
-      Sandbox::Attribute.new(name: name).with_anchor(anchor)
+      Sandbox::Profile::Request::Attribute.new(name: name).with_anchor(anchor)
     end
     it 'adds an anchor' do
-      expected = '{"name":"name","value":"","derivation":"","optional":false,"anchors":[{"type":"SOURCE","value":"anchorValue","sub_type":"","timestamp":1215302400}]}'
+      expected = '{"name":"name","value":"","derivation":"","optional":false,"anchors":[{"type":"SOURCE","value":"anchorValue","sub_type":"","timestamp":1215302400000000}]}'
       expect(attribute.to_json).to eql expected
     end
   end
   describe '.to_json' do
     let :attribute do
-      Sandbox::Attribute.new(
+      Sandbox::Profile::Request::Attribute.new(
         name: name,
         value: value,
         derivation: derivation,
@@ -40,12 +40,12 @@ describe 'Sandbox::Attribute' do
   end
   describe 'self.age_over' do
     it 'creates an age over derivation' do
-      expect(Sandbox::Derivation.age_over(18)).to eql 'age_over:18'
+      expect(Sandbox::Profile::Request::Derivation.age_over(18)).to eql 'age_over:18'
     end
   end
   describe 'self.age_under' do
     it 'creates an age under derivation' do
-      expect(Sandbox::Derivation.age_under(18)).to eql 'age_under:18'
+      expect(Sandbox::Profile::Request::Derivation.age_under(18)).to eql 'age_under:18'
     end
   end
 end
